allow switching of Pose and Tasks-vision model using env variable REACT_APP_MODEL
